SDIT-2832 Refactor - split tests and move to tap package
